Port: removed useless msg->cancelled == 0 checks.
In src/nxt_port_socket.c::nxt_port_read_msg_process() msg->cancelled is
set to 0 and is not touched again.

However there are several checks for it being == 0 which are _always_
true, so remove them.

I'm assuming that this is functioning as intended and that setting
msg->cancelled to 0 is correct.

msg->cancelled was set to 0 unconditionally in commit e227fc9
("Introducing application and port shared memory queues."), so I guess
the now redundant checks were simply missed for removal.